misc/scaffolding/react-flask/backend/app.py
# Flask Application - Single File Backend
# LLM: Implement all models, routes, and logic in this file

from flask import Flask, jsonify, request
from flask_sqlalchemy import SQLAlchemy
from flask_cors import CORS
from functools import wraps
from datetime import datetime, timedelta
import os
import jwt
import bcrypt
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = Flask(__name__)

# CORS config
cors_origins = os.environ.get('CORS_ORIGINS', '').strip()
if cors_origins:
    origins_list = [o.strip() for o in cors_origins.split(',') if o.strip()]
    CORS(app, supports_credentials=True, origins=origins_list)
else:
    CORS(app, supports_credentials=True)

# Database config
app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL', 'sqlite:////app/data/app.db')
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY', 'dev-secret-key')

db = SQLAlchemy(app)

# =============================================================================
# MODELS - LLM: Implement User model and app-specific models
# =============================================================================

# LLM: Implement User model with:
# - id, username, email, password_hash, is_admin, is_active, created_at
# - set_password(password), check_password(password), to_dict() methods

# LLM: Add application-specific models below


# =============================================================================
# AUTH HELPERS - LLM: Implement JWT auth
# =============================================================================

# LLM: Implement token_required decorator
# LLM: Implement admin_required decorator
# LLM: Implement generate_token(user) function


# =============================================================================
# AUTH ROUTES - /api/auth/*
# =============================================================================

# LLM: Implement POST /api/auth/register
# LLM: Implement POST /api/auth/login  
# LLM: Implement GET /api/auth/me


# =============================================================================
# USER ROUTES - /api/* - LLM: Implement user-facing endpoints
# =============================================================================

# LLM: Implement CRUD endpoints for user resources
# Example: GET /api/items, POST /api/items, PUT /api/items/<id>, DELETE /api/items/<id>


# =============================================================================
# ADMIN ROUTES - /api/admin/* - LLM: Implement admin endpoints
# =============================================================================

# LLM: Implement admin endpoints
# Example: GET /api/admin/stats, GET /api/admin/users


# =============================================================================
# HEALTH CHECK
# =============================================================================

@app.route('/api/health')
def health():
    return jsonify({'status': 'healthy'})


# =============================================================================
# INIT
# =============================================================================

def init_app():
    with app.app_context():
        db.create_all()
        logger.info("Database initialized")
        # LLM: Create default admin user here if needed


init_app()

if __name__ == '__main__':
    port = int(os.environ.get('FLASK_RUN_PORT', 5000))
    app.run(host='0.0.0.0', port=port)
